requirements.txt
# Python 3.9 Compatible Requirements
# Core dependencies
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.21.0
plotly>=5.0.0
matplotlib>=3.5.0

# Financial data APIs
yfinance>=0.2.0
mftool>=0.3.0
deprecated>=1.2.0  # Required by mftool
indstocks>=0.1.0  # Requires Python >=3.10, not compatible with Python 3.9
requests>=2.28.0
httpx>=0.24.0
beautifulsoup4>=4.11.0
html5lib>=1.1
lxml>=4.9.0
selenium>=4.10.0  # For web scraping if needed
webdriver-manager>=4.0.0  # For automatic ChromeDriver management

# Database and ORM
supabase>=1.0.0
postgrest>=0.10.0
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# File processing
openpyxl>=3.0.0
xlrd>=2.0.0
python-docx>=0.8.0

# Additional dependencies
typing-extensions>=4.0.0
python-dateutil>=2.8.0
pytz>=2022.1
urllib3>=1.26.0
certifi>=2022.12.7

# AI and Chatbot
openai>=2.0.0  # Updated for latest API
google-generativeai>=0.8.0  # Updated for Gemini 2.5 models
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.0

# Streamlit Cloud specific
gunicorn>=20.1.0
